+++ gems/aws-sdk-chimesdkmeetings/lib/aws-sdk-chimesdkmeetings/client.rb
@@ -945,7 +945,7 @@ module Aws::ChimeSDKMeetings
         params: params,
         config: config)
       context[:gem_name] = 'aws-sdk-chimesdkmeetings'
-      context[:gem_version] = '1.9.0'
+      context[:gem_version] = '1.10.0'
       Seahorse::Client::Request.new(handlers, context)
     end