+++ gems/aws-sdk-core/lib/aws-sdk-sso.rb
@@ -50,6 +50,6 @@ require_relative 'aws-sdk-sso/customizations'
 # @!group service
 module Aws::SSO
 
-  GEM_VERSION = '3.130.1'
+  GEM_VERSION = '3.130.2'
 
 end