+++ gems/aws-sdk-core/lib/aws-sdk-sso/client.rb
@@ -545,7 +545,7 @@ module Aws::SSO
         params: params,
         config: config)
       context[:gem_name] = 'aws-sdk-core'
-      context[:gem_version] = '3.130.1'
+      context[:gem_version] = '3.130.2'
       Seahorse::Client::Request.new(handlers, context)
     end