+++ gems/aws-sdk-core/CHANGELOG.md
@@ -1,6 +1,9 @@
 Unreleased Changes
 ------------------
 
+3.130.2 (2022-04-22)
+------------------
+
 * Issue - Don't pass `:before_refresh` to Client constructors in RefreshingCredential implementations (#2690).
 
 3.130.1 (2022-04-12)